requirements.txt
absl-py==2.1.0
authlib==1.3.1
ansible==9.3.0
ansible-core==2.16.4
apache-libcloud==3.2.0
argcomplete==1.8.1
astunparse==1.6.3
Babel==2.8.0
bcrypt==3.2.0
blinker==1.7.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
cryptography==3.4.8
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
mysql-connector==2.2.9
mysql-connector-python==8.3.0
gunicorn==20.1.0
flask-session==0.4.0
python-dotenv==0.21.0
